pyproject.toml
[project]
name = "adk-rfp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.15.1",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-aiplatform>=1.117.0",
    "google-genai>=1.39.1",
    "ipykernel>=6.30.1",
    "markdown>=3.9",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "pypandoc>=1.15",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
]
